# openapi/esim_go_schema_v2_5.yaml
openapi: 3.0.1
info:
  title: eSIM-Go API
  version: 2.5.0
  description: |
    To access the eSIMGo API, you need to authenticate your requests using an API key.
    Here's a brief overview of the authentication process:
      1. Assuming that you have created account already:
      - Log into your eSIMGo account at https://sso.esim-go.com/login.
      - Navigate to Account Settings -> API Details to find your API key.
      - Keep your API key secure and avoid sharing it with others.
      2. API Key Usage:
      - Include your API key in the header of all eSIMGo API requests.
      - Use the header key 'X-API-KEY' with your API key as the value.
      3. Security Scheme:
      - The eSIMGo API uses an HTTP security scheme type for authentication.
      4. Authorization Types:
      - The API supports 'apiKeyAuth' authorization type.
servers:
  - url: https://api.esim-go.com/v2.5
    description: eSIMGo API
...
paths:
...
  /orders:
    get:
      summary: List orders
      description: |
        Get details on all previous orders, including total cost and contents. Response data is paginated.
      tags:
        - Orders
      parameters:
        - name: Accept
          in: header
          schema:
            type: string
          example: application/json
        - name: includeIccids
          in: query
          schema:
            type: boolean
          description: |
            Set to true to include eSIM data (ICCID, Matching ID and SMDP Address) in the response, and an ICCIDs array.
          example: "true"
        - name: page
          in: query
          schema:
            type: integer
          description: |
            Page number to return.
          example: "1"
        - name: limit
          in: query
          schema:
            type: integer
          description: |
            Number of results to return per page.
          example: "10"
        - name: createdAt
          in: query
          schema:
            type: string
          description: |
            Specifies the date range for filtering orders. This parameter has a 'lte:' prefix to specify the end date. For example, to query orders from March 1, 2024, to March 31, 2024, use the following format: `createdAt=gte:2024-03-01T00:00:00.000Z&createdAt=lte:2024-03-31T23:59:59.999Z`.
          example: <dateTime>
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponseTransaction"
        "400":
          description: Bad Request - returned when request format is not accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "403":
          description: Unauthorised - returned when the api token is invalid; the user does not have any available Bundles left or the ICCID is not accessible by the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "503":
          description: Processing - Please come back later or use the *Retry-After* (seconds) header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
      x-codeSamples:
        - lang: shell
          label: cURL
          source: |-
            curl 'https://api.esim-go.com/v2.5/orders' \
            -H 'X-API-Key: $API_KEY' \
            -H 'Content-Type: application/json'
        - lang: javascript
          label: Node
          source: |
            const axios = require('axios');

            let config = {
              method: 'get',
              maxBodyLength: Infinity,
              url: 'https://api.esim-go.com/v2.5/orders',
              headers: { 
                'X-API-Key': '$API_KEY', 
                'Content-Type': 'application/json'
              }
            };

            axios.request(config)
            .then((response) => {
              console.log(JSON.stringify(response.data));
            })
            .catch((error) => {
              console.log(error);
            });
        - lang: go
          label: Go
          source: |-
            package main

            import (
              "fmt"
              "net/http"
              "io"
            )

            func main() {

              url := "https://api.esim-go.com/v2.5/orders"
              method := "GET"

              client := &http.Client {
              }
              req, err := http.NewRequest(method, url, nil)

              if err != nil {
                fmt.Println(err)
                return
              }
              req.Header.Add("X-API-Key", "$API_KEY")
              req.Header.Add("Content-Type", "application/json")

              res, err := client.Do(req)
              if err != nil {
                fmt.Println(err)
                return
              }
              defer res.Body.Close()

              body, err := io.ReadAll(res.Body)
              if err != nil {
                fmt.Println(err)
                return
              }
              fmt.Println(string(body))
            }
        - lang: python
          label: Python
          source: |
            import requests
            import json

            url = "https://api.esim-go.com/v2.5/orders"

            payload = {}
            headers = {
              'X-API-Key': '$API_KEY',
              'Content-Type': 'application/json'
            }

            response = requests.request("GET", url, headers=headers, data=payload)

            print(response.text)
        - lang: csharp
          label: C#
          source: |
            var client = new HttpClient();
            var request = new HttpRequestMessage(HttpMethod.Get, "https://api.esim-go.com/v2.5/orders");
            request.Headers.Add("X-API-Key", "$API_KEY");
            var content = new StringContent(string.Empty);
            content.Headers.ContentType = new MediaTypeHeaderValue("application/json");
            request.Content = content;
            var response = await client.SendAsync(request);
            response.EnsureSuccessStatusCode();
            Console.WriteLine(await response.Content.ReadAsStringAsync());
        - lang: java
          label: Java
          source: |
            Unirest.setTimeouts(0, 0);
            HttpResponse<String> response = Unirest.get("https://api.esim-go.com/v2.5/orders")
              .header("X-API-Key", "$API_KEY")
              .header("Content-Type", "application/json")
              .asString();
    post:
      summary: Create orders
      description: |
        Orders can be validated and processed using this endpoint.

        Total will be deducted from your Organisation's balance.

        ## Auto-assign

        By specifying ICCID(s) of eSIM(s) belonging to your Organisation and setting 'assign' to true, a bundle can be automatically assigned to eSIM(s).

        If specified ICCID(s) of eSIM(s) and the specified bundle(s) are not
        compatible, and ‘allowReassign’ set to true, the bundle(s) will be
        assigned to new ICCID(s).

        Bundle assignments to an eSIM are usually instant but please allow forup to 10 minutes for the bundle to fully process. While the bundle is processing the eSIM can be successfully installed and the eSIM will register onto a network if within coverage.

        The Bundle Status can be checked through [Get the Status of a bundle assigned to an eSIM](/api/v2_4/operations/esimsiccidbundlesname/get/)

        ## Branding Profile ID:

        This is optional and only used if you have multiple branding profiles. It allows you to generate an eSIM with a specific branding profile instead of the default (first) one. To find the Profile ID:

        1. Log in to your portal account

        2. Navigate to "eSIM Branding"

        3. Locate the desired branding profile

        4. Note the Profile ID associated with it

        If not specified, the default (first) branding profile will be used.

        ## Usage:

        - eSIM ICCIDs can only be specified if assign is set to true
        - If assign is set to true, but no ICCIDs are provided, bundles are
          assigned to new eSIMs
        - If ICCIDs are provided, quantity is required to match the number of
          ICCIDs for each bundle
        - If quantity is specified and assign is set to false, the quantity of
          that bundle is purchased into inventory
        - If new bundles and specified eSIM ICCIDs are not compatible and
        ‘allowReassign’ set to true, bundles will be assigned to new ICCIDs
        - To use profile id ensure that assign is set to true
      tags:
        - Orders
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequest"
            example:
              type: transaction
              assign: true
              order:
                - type: bundle
                  quantity: 3
                  item: esim_10GB_30D_IM_U
                  allowReassign: false
                - type: bundle
                  quantity: 5
                  item: esim_10GB_30D_IM_U
                  iccids:
                    - "8901234567890123456"
                    - "8901234567890123457"
                  allowReassign: true
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/OrderResponseValidate"
                  - $ref: "#/components/schemas/OrderResponseTransaction"
              examples:
                validateResponse:
                  summary: Example of a validation response
                  value:
                    $ref: "#/components/schemas/OrderResponseValidate/example"
                transactionResponse:
                  summary: Example of a transaction response
                  value:
                    $ref: "#/components/schemas/OrderResponseTransaction/example"
        "400":
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/Message"
              example: ""
        "401":
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/Message"
              example: ""
        "503":
          description: Service Unavailable
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/Message"
              example: ""
      x-codeSamples:
        - lang: shell
          label: cURL
          source: |-
            curl 'https://api.esim-go.com/v2.5/orders' \
            -H 'X-API-Key: $API_KEY' \
            -H 'Content-Type: application/json' \
            -d '{"$ref":"#/components/schemas/OrderRequest"}'
        - lang: javascript
          label: Node
          source: |
            const axios = require('axios');
            let data = JSON.stringify({
              "$ref": "#/components/schemas/OrderRequest"
            });

            let config = {
              method: 'post',
              maxBodyLength: Infinity,
              url: 'https://api.esim-go.com/v2.5/orders',
              headers: { 
                'X-API-Key': '$API_KEY', 
                'Content-Type': 'application/json'
              },
              data : data
            };

            axios.request(config)
            .then((response) => {
              console.log(JSON.stringify(response.data));
            })
            .catch((error) => {
              console.log(error);
            });
        - lang: go
          label: Go
          source: |-
            package main

            import (
              "fmt"
              "strings"
              "net/http"
              "io"
            )

            func main() {

              url := "https://api.esim-go.com/v2.5/orders"
              method := "POST"

              payload := strings.NewReader(`{"$ref":"#/components/schemas/OrderRequest"}`)

              client := &http.Client {
              }
              req, err := http.NewRequest(method, url, payload)

              if err != nil {
                fmt.Println(err)
                return
              }
              req.Header.Add("X-API-Key", "$API_KEY")
              req.Header.Add("Content-Type", "application/json")

              res, err := client.Do(req)
              if err != nil {
                fmt.Println(err)
                return
              }
              defer res.Body.Close()

              body, err := io.ReadAll(res.Body)
              if err != nil {
                fmt.Println(err)
                return
              }
              fmt.Println(string(body))
            }
        - lang: python
          label: Python
          source: |
            import requests
            import json

            url = "https://api.esim-go.com/v2.5/orders"

            payload = json.dumps({
              "$ref": "#/components/schemas/OrderRequest"
            })
            headers = {
              'X-API-Key': '$API_KEY',
              'Content-Type': 'application/json'
            }

            response = requests.request("POST", url, headers=headers, data=payload)

            print(response.text)
        - lang: csharp
          label: C#
          source: |
            var client = new HttpClient();
            var request = new HttpRequestMessage(HttpMethod.Post, "https://api.esim-go.com/v2.5/orders");
            request.Headers.Add("X-API-Key", "$API_KEY");
            var content = new StringContent("{\"$ref\":\"#/components/schemas/OrderRequest\"}", null, "application/json");
            request.Content = content;
            var response = await client.SendAsync(request);
            response.EnsureSuccessStatusCode();
            Console.WriteLine(await response.Content.ReadAsStringAsync());
        - lang: java
          label: Java
          source: |
            Unirest.setTimeouts(0, 0);
            HttpResponse<String> response = Unirest.post("https://api.esim-go.com/v2.5/orders")
              .header("X-API-Key", "$API_KEY")
              .header("Content-Type", "application/json")
              .body("{\"$ref\":\"#/components/schemas/OrderRequest\"}")
              .asString();
  /orders/{orderReference}:
    get:
      summary: Get order detail
      description: |
        Get details on an order, including total cost and contents
      tags:
        - Orders
      parameters:
        - name: Accept
          in: header
          schema:
            type: string
          example: application/json
        - name: orderReference
          in: path
          schema:
            type: string
          required: true
          description: (Required) Reference for your order
          example: <string>
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponseTransaction"
              example:
                order:
                  - esims:
                      - iccid: "8912345678901234567"
                        matchingId: AB-12C3DE-4FGHIJ5
                        smdpAddress: http://rsp.mockprovider.com
                    type: bundle
                    item: esim_5GB_30D_EU_V3
                    iccids:
                      - "8912345678901234567"
                    quantity: 1
                    subTotal: 9.99
                    pricePerUnit: 9.99
                    AllowReassign: true
                total: 9.99
                currency: USD
                status: completed
                statusMessage: "Order completed: 1 eSIMs assigned"
                orderReference: 1a2b3c4d-5e6f-7g8h-9i10-11j12k13l14m
                createdDate: "2023-06-15T10:30:45.123456789Z"
                assigned: true
        "400":
          description: Bad Request - returned when request format is not accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "403":
          description: Unauthorised - returned when the api token is invalid; the user does not have any available Bundles left or the ICCID is not accessible by the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "503":
          description: Processing - Please come back later or use the *Retry-After* (seconds) header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
      x-codeSamples:
        - lang: shell
          label: cURL
          source: |-
            curl -g 'https://api.esim-go.com/v2.5/orders/{orderReference}' \
            -H 'X-API-Key: $API_KEY' \
            -H 'Content-Type: application/json'
        - lang: javascript
          label: Node
          source: |
            const axios = require('axios');

            let config = {
              method: 'get',
              maxBodyLength: Infinity,
              url: 'https://api.esim-go.com/v2.5/orders/{orderReference}',
              headers: { 
                'X-API-Key': '$API_KEY', 
                'Content-Type': 'application/json'
              }
            };

            axios.request(config)
            .then((response) => {
              console.log(JSON.stringify(response.data));
            })
            .catch((error) => {
              console.log(error);
            });
        - lang: go
          label: Go
          source: |-
            package main

            import (
              "fmt"
              "net/http"
              "io"
            )

            func main() {

              url := "https://api.esim-go.com/v2.5/orders/{orderReference}"
              method := "GET"

              client := &http.Client {
              }
              req, err := http.NewRequest(method, url, nil)

              if err != nil {
                fmt.Println(err)
                return
              }
              req.Header.Add("X-API-Key", "$API_KEY")
              req.Header.Add("Content-Type", "application/json")

              res, err := client.Do(req)
              if err != nil {
                fmt.Println(err)
                return
              }
              defer res.Body.Close()

              body, err := io.ReadAll(res.Body)
              if err != nil {
                fmt.Println(err)
                return
              }
              fmt.Println(string(body))
            }
        - lang: python
          label: Python
          source: |
            import requests
            import json

            url = "https://api.esim-go.com/v2.5/orders/{orderReference}"

            payload = {}
            headers = {
              'X-API-Key': '$API_KEY',
              'Content-Type': 'application/json'
            }

            response = requests.request("GET", url, headers=headers, data=payload)

            print(response.text)
        - lang: csharp
          label: C#
          source: |
            var client = new HttpClient();
            var request = new HttpRequestMessage(HttpMethod.Get, "https://api.esim-go.com/v2.5/orders/{orderReference}");
            request.Headers.Add("X-API-Key", "$API_KEY");
            var content = new StringContent(string.Empty);
            content.Headers.ContentType = new MediaTypeHeaderValue("application/json");
            request.Content = content;
            var response = await client.SendAsync(request);
            response.EnsureSuccessStatusCode();
            Console.WriteLine(await response.Content.ReadAsStringAsync());
        - lang: java
          label: Java
          source: |
            Unirest.setTimeouts(0, 0);
            HttpResponse<String> response = Unirest.get("https://api.esim-go.com/v2.5/orders/{orderReference}")
              .header("X-API-Key", "$API_KEY")
              .header("Content-Type", "application/json")
              .asString();
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      x-displayName: API Key in Header
      description: |
        Your API Key can passed to the eSIMGo API by setting a Header with the key `X-API-Key` in your HTTP call.

        e.g.

        ```
        curl -H 'X-API-Key: $API_KEY' https://api.esim-go.com/v2.3/esims/$ICCID/bundles
        ```
  schemas:
...
    Message:
      type: object
      properties:
        message:
          description: status of api functions which do not return data
          type: string
...
    OrderRequest:
      type: object
      properties:
        type:
          type: string
          enum:
            - validate
            - transaction
        assign:
          type: boolean
          description: To specify if bundle(s) to be assigned to eSIM(s)
        order:
          type: array
          items:
            $ref: "#/components/schemas/BundleOrder"
        profileID:
          type: string
          description: Optional identifier for selecting a specific eSIM branding profile when multiple profiles exist.
    BundleOrder:
      type: object
      properties:
        type:
          type: string
          enum:
            - bundle
        quantity:
          type: integer
          description: Quantity of the bundles to purchase. If you purchase several different bundles, specify the quantity of each item separately.
        item:
          type: string
          description: Name of the bundle
        iccids:
          type: array
          description: ICCID of the eSIM to assign bundle to (quantity of items and number of ICCIDs should match)
          items:
            type: string
        allowReassign:
          type: boolean
          description: Allow a new eSIM to be provided if the bundle is not compatible with the eSIM profile
    OrderResponseValidate:
      type: object
      properties:
        order:
          type: array
          description: Order type [ validate ]
          items:
            type: object
            properties:
              type:
                type: string
                description: Item type [ bundle ]
              item:
                type: string
                description: Name of the bundle
              quantity:
                type: integer
                description: Quantity of items
              subTotal:
                type: number
                format: float
                description: Cost of a bundle multiplied by its quantity
              pricePerUnit:
                type: number
                format: float
                description: Price per each item
              AllowReassign:
                type: boolean
                description: Allow a new eSIM to be provided if the bundle is not compatible with the eSIM profile
        total:
          type: number
          format: float
          description: Total price
        valid:
          type: boolean
          description: Indicates if request is valid
        currency:
          type: string
          description: Currency of the transaction
        createdDate:
          type: string
          format: date-time
          description: Data and time of order creation
        assigned:
          type: boolean
          description: Indicates if bundle was assigned to eSIM
      example:
        order:
          - type: bundle
            item: esim_1GB_7D_AX_V2
            quantity: 1
            subTotal: 2.28
            pricePerUnit: 2.28
            AllowReassign: false
        total: 2.28
        valid: true
        currency: USD
        createdDate: "2025-03-28T14:47:26.153875019Z"
        assigned: true
    OrderResponseTransaction:
      type: object
      properties:
        order:
          type: array
          description: Order type [ transaction ]
          items:
            type: object
            properties:
              esims:
                type: array
                items:
                  type: object
                  properties:
                    iccid:
                      type: string
                      description: ICCID of eSIM new bundle was assigned to
                    matchingId:
                      type: string
                      description: SMDP+ ID (used in eSIM activation)
                    smdpAddress:
                      type: string
                      description: SMDP+ Address (used in eSIM activation)
              type:
                type: string
                description: Item type [ bundle ]
              item:
                type: string
                description: Name of the bundle
              iccids:
                type: array
                description: ICCID of eSIM new bundle was assigned to
                items:
                  type: string
              quantity:
                type: integer
                description: Quantity of items
              subTotal:
                type: number
                format: float
                description: Cost of a bundle multiplied by its quantity
              pricePerUnit:
                type: number
                format: float
                description: Price per each item
              AllowReassign:
                type: boolean
                description: Allow a new eSIM to be provided if the bundle is not compatible with the eSIM profile
        total:
          type: number
          format: float
          description: Total price
        currency:
          type: string
          description: Currency of the transaction
        status:
          type: string
          description: Status of the order
        statusMessage:
          type: string
          description: Status message
        orderReference:
          type: string
          description: Order reference
        createdDate:
          type: string
          format: date-time
          description: Data and time of order creation
        assigned:
          type: boolean
          description: Indicates if bundle was assigned to eSIM
        sourceIP:
          type: string
          description: Source IP of the order
        runningBalance:
          type: string
          description: Current balance at a time when the order was placed
      example:
        order:
          - type: bundle
            item: esim_data_plan_2
            quantity: 1
            subTotal: 9.99
            pricePerUnit: 9.99
            AllowReassign: false
            esims:
              - iccid: "8900000000000000001"
                matchingId: MATCH-001
                smdpAddress: http://smdp.example.com
            iccids:
              - "8900000000000000001"
        total: 9.99
        currency: USD
        status: completed
        statusMessage: "Order completed: 1 eSIM assigned"
        orderReference: order-ref-002
        createdDate: "2023-06-14T15:45:00Z"
        assigned: true
...